Automated time series download
A rough implementation of automated time series data download for hindcasting events. Specify in the code the start and end dates of the event, the spin-up time for the model, and interval hours within a day, and then the time period will be split into (the least number of) requests. The data would be then downloaded.